requirements.txt
matplotlib==3.10.7
numpy==1.25.0
opencv_contrib_python==4.10.0.84
opencv_python_headless==4.10.0.84
picamera2==0.3.31
Pillow==12.0.0
pytesseract==0.3.13
tensorflow==2.20.0
